Split message handling code to separated file
