+++ .gitignore
@@ -6,4 +6,5 @@ synapsebeacon
 synapsevalidator
 synapsep2p
 integration_tests
+integration
 *.exe
